refactor: move realtime types to api folder
